code/src/chainlit/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
argostranslate==1.9.6
asyncer==0.0.7
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
browser-use==0.1.38
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chainlit==2.2.0
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.1
ctranslate2==4.5.0
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==8.4.2
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==7.5.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0 
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
googletrans==4.0.0rc1
greenlet==3.1.1
gretel-client==0.26.0
groq==0.18.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==3.2.0
holidays==0.69
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
hyperframe==5.2.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.19
langchain-anthropic==0.3.3
langchain-community==0.3.18
langchain-core==0.3.37
langchain-google-genai==2.0.10
langchain-groq==0.2.4
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.11
Lazify==0.4.0
literalai==0.1.103
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.5.0
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
mtranslate==1.8
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.7
onnxruntime==1.19.2
openai==1.63.2
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orderly-set==5.3.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pefile==2023.2.7
pillow==11.1.0
playwright==1.50.0
posthog==3.16.0
primp==0.14.0
propcache==0.3.0
prophet==1.1.6
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pymongo==4.11.1
pyparsing==3.2.1
pypdf==5.3.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.2
python-socketio==5.12.1
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
sacremoses==0.0.53
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
setuptools==75.8.2
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sse-starlette==2.2.1
stanio==0.5.1
stanza==1.1.1
starlette==0.41.3
statsmodels==0.14.4
sympy==1.13.3
syncer==2.0.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.16.23
unstructured-client==0.25.9
uptrace==1.29.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
webencodings==0.5.1
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.17
